requirements.txt
numpy
pillow
gymnasium
streamlit
matplotlib
stable-baselines3==2.3.2